refactor: Renamed functions for create-file-path-maps (ember-addon)
