+++ src/migration/ember-addon/steps/create-file-path-maps.ts
@@ -1,9 +1,9 @@
 import type { FilePathMap } from '@codemod-utils/files';
 
 import type { Options } from '../../../types/index.js';
-import { migrationStrategyForAddonFolder } from './create-file-path-maps/addon/index.js';
-import { migrationStrategyForAppFolder } from './create-file-path-maps/app/index.js';
-import { migrationStrategyForTestsFolder } from './create-file-path-maps/tests/index.js';
+import { mapAddonFolder } from './create-file-path-maps/addon/index.js';
+import { mapAppFolder } from './create-file-path-maps/app/index.js';
+import { mapTestsFolder } from './create-file-path-maps/tests/index.js';
 
 type MigrationStrategies = {
   addon: FilePathMap;
@@ -12,13 +12,9 @@ type MigrationStrategies = {
 };
 
 export function createFilePathMaps(options: Options): MigrationStrategies {
-  const filePathMapAddon = migrationStrategyForAddonFolder(options);
-  const filePathMapApp = migrationStrategyForAppFolder(options);
-  const filePathMapTests = migrationStrategyForTestsFolder(options);
-
   return {
-    addon: filePathMapAddon,
-    app: filePathMapApp,
-    tests: filePathMapTests,
+    addon: mapAddonFolder(options),
+    app: mapAppFolder(options),
+    tests: mapTestsFolder(options),
   };
 }
