+++ src/migration/ember-addon/steps/create-file-path-maps/app/map-component-templates.ts
@@ -5,26 +5,21 @@ import type {
   Options,
 } from '../../../../../types/index.js';
 
-export function migrationStrategyForComponents(
-  options: Options,
-): FilePathMapEntries {
+export function mapComponentTemplates(options: Options): FilePathMapEntries {
   const { projectRoot } = options;
 
-  const filePaths = findFiles(
-    'tests/integration/components/**/component-test.{js,ts}',
-    {
-      projectRoot,
-    },
-  );
+  const filePaths = findFiles('app/components/**/template.js', {
+    projectRoot,
+  });
 
   return filePaths.map((oldFilePath) => {
     const newFilePath = renamePathByFile(oldFilePath, {
       find: {
-        directory: 'tests/integration/components',
-        file: 'component-test',
+        directory: 'app/components',
+        file: 'template',
       },
       replace: (key: string) => {
-        return `tests/integration/components/${key}-test`;
+        return `app/components/${key}`;
       },
     });
