+++ src/migration/ember-addon/steps/create-file-path-maps/tests/index.ts
@@ -1,6 +1,6 @@
 import type { FilePathMap, Options } from '../../../../../types/index.js';
-import { migrationStrategyForComponents } from './components.js';
+import { mapComponents } from './map-components.js';
 
-export function migrationStrategyForTestsFolder(options: Options): FilePathMap {
-  return new Map([...migrationStrategyForComponents(options)]);
+export function mapTestsFolder(options: Options): FilePathMap {
+  return new Map([...mapComponents(options)]);
 }
